pyproject.toml
[tool.poetry]

# Change this to the name of your own extension
name = "skybrush-ext-template"
version = "1.0.0"
description = "Template project for Skybrush Server extension modules"
authors = ["Tamas Nepusz <ntamas@gmail.com>"]
license = "MIT"
packages = [
  # Change this to the name of the folder in src/ containing your extension
  { include = "skybrush_ext_template", from = "src" }
]

[[tool.poetry.source]]
name = "fury"
url = "https://pypi.fury.io/skybrush/"
secondary = true

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
flockwave-server = { git = "https://github.com/skybrush-io/skybrush-server", branch = "main" }

[tool.poetry.plugins]

[tool.poetry.plugins."skybrush.server.ext"]
# Change the key below to the name of the extension that you want to use in
# configuration files, and the value to the name of the Python module that
# contains the extension code
template = "skybrush_ext_template"

[tool.poetry.scripts]
skybrushd = "flockwave.server.launcher:start"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
